Rename `api` to `core`
