--- docs/pymagicc.api.rst
@@ -1,8 +1,0 @@
-.. include:: _custom_rst_shortcuts.rst
-
-pymagicc.api
-------------
-
-.. automodule:: pymagicc.api
-
-
