+++ docs/pymagicc.core.rst
@@ -1,0 +1,6 @@
+.. include:: _custom_rst_shortcuts.rst
+
+pymagicc.core
+-------------
+
+.. automodule:: pymagicc.core
