+++ pymagicc/core.py
