+++ tests/test_api.py
@@ -11,7 +11,7 @@ import pandas as pd
 import f90nml
 
 
-from pymagicc.api import MAGICCBase, MAGICC6, MAGICC7, config, _clean_value
+from pymagicc.core import MAGICCBase, MAGICC6, MAGICC7, config, _clean_value
 from pymagicc.io import MAGICCData
 from .test_config import config_override  #  noqa
